requirements.txt
gradio>=4.0.0
flask>=2.3.0
python-docx>=0.8.11
openai>=1.0.0
requests>=2.31.0
werkzeug>=2.3.0
python-dotenv>=1.0.0